+++ packages/e2e-tests/scripts/deployAndDeposit.ts
@@ -43,14 +43,26 @@ const TEST_CANONICAL_NOCTURNE_ADDRS: CanonAddress[] = [
   },
 ];
 
+const SUBTREE_BATCH_FILLER: string | undefined = process.env.SUBTREE_BATCH_FILLER;
+
 (async () => {
   console.log("deploying contracts with dummy proxy admin...");
   const provider = new ethers.providers.JsonRpcProvider(HH_URL);
   const chainId = BigInt((await provider.getNetwork()).chainId);
   const [deployerEoa] = KEYS_TO_WALLETS(provider);
+
+  const subtreeBatchFillers = [deployerEoa.address];
+  if (SUBTREE_BATCH_FILLER) {
+    const eoa = new ethers.Wallet(SUBTREE_BATCH_FILLER);
+    subtreeBatchFillers.push(eoa.address);
+  }
+
+  console.log("subtree batch fillers:", subtreeBatchFillers);
+
   const { depositManagerProxy, walletProxy } =
     await deployContractsWithDummyAdmins(deployerEoa, {
-      screeners: [deployerEoa.address], // TODO: remove this once we have real deposit-screener
+      screeners: [deployerEoa.address], // TODO: remove this once we have real deposit-screener,
+      subtreeBatchFillers,
     });
   const wallet = Wallet__factory.connect(walletProxy.proxy, deployerEoa);
   const depositManager = DepositManager__factory.connect(