+++ packages/site/src/pages/index.tsx
@@ -35,7 +35,7 @@ import { VAULT_CONTRACT_ADDRESS, WALLET_CONTRACT_ADDRESS } from "../config";
 import { TxModal } from "../components/TxModal";
 
 const ERC20_ID = 0n;
-const TOKEN_ADDRESS = "0x510C6297cC30A058F41eb4AF1BFC9953EaD8b577";
+const TOKEN_ADDRESS = "0x4e3df2073bf4b43B9944b8e5A463b1E185D6448C";
 
 const Container = styled.div`
   display: flex;